Threw in the unity project for testing rope
All future changes to the rope I'm going to work on in that project
file.  Don't mess with it unless you've already told me you're going to
be messing around with the web-- it probably won't get used much.
Justin you're all good to get the splines working inside of that file if
you want, in fact I'll need to get the equation of the line that you're
drawing in there at some point anyway.