Added Moving Platforms
They're called carriages.  Technically you can apply the script to
anything as well, but I set it up via an example object, so you can just
use that and edit it as needed.  Messing with the collision box will
have no negative effects, just don't touch the script or add a
rigidbody2D to it.